fix_hw

+++ navy-apps/libs/libos/src/nanos.c
@@ -35,7 +35,7 @@ int _open(const char *path, int flags, mode_t mode) {
 }
 
 int _write(int fd, void *buf, size_t count){
-  int ret = _syscall_(SYS_write, fd, (int)buf, count);
+  int ret = _syscall_(SYS_write, fd, buf, count);
   return ret;
 }
